refactor(plugins): move space tab emulator component to separate file
