+++ packages/sdk/src/emulator/HostEmulator.styled.ts
@@ -38,13 +38,3 @@ export const SpaceContent = styled.div`
   border: 1px solid grey;
   margin: 10px;
 `;
-
-export const SpaceTopBar = styled.div`
-  display: flex;
-  height: 50px;
-  border: 1px solid grey;
-  border-radius: 4px;
-  margin: 10px;
-  padding: 10px;
-  gap: 10px;
-`;
