+++ packages/sdk/src/emulator/components/SpaceTabEmulator/SpaceTabEmulator.styled.tsx
@@ -1,0 +1,11 @@
+import styled from 'styled-components';
+
+export const SpaceTopBar = styled.div`
+  display: flex;
+  height: 50px;
+  border: 1px solid grey;
+  border-radius: 4px;
+  margin: 10px;
+  padding: 10px;
+  gap: 10px;
+`;
