+++ packages/sdk/src/emulator/components/SpaceTabEmulator/index.ts
@@ -1,0 +1,1 @@
+export * from './SpaceTabEmulator';
